+++ verik-compiler/src/main/kotlin/io/verik/compiler/ast/element/declaration/kt/EKtFunction.kt
@@ -67,7 +67,7 @@ class EKtFunction(
     }
 
     override fun accept(visitor: Visitor) {
-        return visitor.visitKtFunction(this)
+        visitor.visitKtFunction(this)
     }
 
     override fun acceptChildren(visitor: TreeVisitor) {