+++ verik-compiler/src/main/kotlin/io/verik/compiler/ast/element/declaration/sv/ECoverGroup.kt
@@ -14,18 +14,18 @@
  * limitations under the License.
  */
 
-package io.verik.compiler.ast.element.declaration.kt
+package io.verik.compiler.ast.element.declaration.sv
 
 import io.verik.compiler.ast.common.Type
 import io.verik.compiler.ast.element.declaration.common.EAbstractContainerClass
 import io.verik.compiler.ast.element.declaration.common.EDeclaration
 import io.verik.compiler.ast.element.declaration.common.ETypeParameter
 import io.verik.compiler.ast.property.AnnotationEntry
-import io.verik.compiler.common.TreeVisitor
 import io.verik.compiler.common.Visitor
 import io.verik.compiler.message.SourceLocation
+import io.verik.compiler.target.common.Target
 
-class EKtClass(
+class ECoverGroup(
     override val location: SourceLocation,
     override val bodyStartLocation: SourceLocation,
     override val bodyEndLocation: SourceLocation,
@@ -33,51 +33,18 @@ class EKtClass(
     override var type: Type,
     override var annotationEntries: List<AnnotationEntry>,
     override var documentationLines: List<String>?,
-    override var superType: Type,
     override var typeParameters: ArrayList<ETypeParameter>,
-    override var declarations: ArrayList<EDeclaration>,
-    var primaryConstructor: EPrimaryConstructor?,
-    var isEnum: Boolean,
-    var isObject: Boolean
+    override var declarations: ArrayList<EDeclaration>
 ) : EAbstractContainerClass() {
+
+    override var superType = Target.C_Void.toType()
 
     init {
         typeParameters.forEach { it.parent = this }
         declarations.forEach { it.parent = this }
-        primaryConstructor?.parent = this
-    }
-
-    fun fill(
-        type: Type,
-        annotationEntries: List<AnnotationEntry>,
-        documentationLines: List<String>?,
-        superType: Type,
-        typeParameters: List<ETypeParameter>,
-        declarations: List<EDeclaration>,
-        primaryConstructor: EPrimaryConstructor?,
-        isEnum: Boolean,
-        isObject: Boolean
-    ) {
-        typeParameters.forEach { it.parent = this }
-        declarations.forEach { it.parent = this }
-        primaryConstructor?.parent = this
-        this.type = type
-        this.annotationEntries = annotationEntries
-        this.documentationLines = documentationLines
-        this.superType = superType
-        this.typeParameters = ArrayList(typeParameters)
-        this.declarations = ArrayList(declarations)
-        this.primaryConstructor = primaryConstructor
-        this.isEnum = isEnum
-        this.isObject = isObject
     }
 
     override fun accept(visitor: Visitor) {
-        visitor.visitKtClass(this)
-    }
-
-    override fun acceptChildren(visitor: TreeVisitor) {
-        super.acceptChildren(visitor)
-        primaryConstructor?.accept(visitor)
+        visitor.visitCoverGroup(this)
     }
 }