+++ verik-compiler/src/main/kotlin/io/verik/compiler/ast/element/declaration/sv/EModule.kt
@@ -40,6 +40,6 @@ class EModule(
     }
 
     override fun accept(visitor: Visitor) {
-        return visitor.visitModule(this)
+        visitor.visitModule(this)
     }
 }