+++ verik-compiler/src/main/kotlin/io/verik/compiler/ast/element/declaration/sv/ESvFunction.kt
@@ -43,6 +43,6 @@ class ESvFunction(
     }
 
     override fun accept(visitor: Visitor) {
-        return visitor.visitSvFunction(this)
+        visitor.visitSvFunction(this)
     }
 }