+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## [Unreleased]
 ### Compiler
 - Support `Double` data type.
-- Add `monitor` and `strobe` functions.
+- Support `monitor` and `strobe` functions.
+- Support `CoverGroup`, `CoverPoint`, and `CoverCross`.
 
 ## [0.1.14]
 ### Importer